requirements.txt
pytorch
transformers
librosa
numpy
scipy
scikit-learn
matplotlib
seaborn
flask
speech_recognition
pydub
